refactor: 199 - break down CreateEditInvoiceForm into smaller components
